+++ src/generic.rs
@@ -1,6 +1,6 @@
 use rand_core::CryptoRng;
 
-use crate::SieveFactory;
+use crate::hazmat::SieveFactory;
 
 /// Sieves through the results of `sieve_factory` and returns the first item for which `predicate` is `true`.
 ///
@@ -37,7 +37,7 @@ mod tests {
     use rand_core::{CryptoRng, OsRng, TryRngCore};
 
     use super::sieve_and_find;
-    use crate::SieveFactory;
+    use crate::hazmat::SieveFactory;
 
     #[test]
     fn test_exhaustable_sieve_factory() {